requirements.txt
accelerate==0.21.0
aiofiles==23.1.0
aiohttp==3.8.5
aiosignal==1.3.1
anyio==3.7.1
async-timeout==4.0.2
asyncer==0.0.2
attrs==23.1.0
auth0-python==4.4.0
backoff==2.2.1
beautifulsoup4==4.12.2
bidict==0.22.1
certifi==2023.7.22
cffi==1.15.1
chainlit==0.6.2
charset-normalizer==3.2.0
click==8.1.6
cohere==4.19.2
colorama==0.4.6
cryptography==41.0.3
dataclasses-json==0.5.14
datasets==2.14.4
Deprecated==1.2.14
dill==0.3.7
dnspython==2.4.1
evaluate==0.4.0
exceptiongroup==1.1.2
fastapi==0.97.0
fastapi-socketio==0.0.10
fastavro==1.8.2
filelock==3.12.2
filetype==1.2.0
frozenlist==1.4.0
fsspec==2023.6.0
googleapis-common-protos==1.60.0
greenlet==2.0.2
grpcio==1.56.2
h11==0.14.0
httpcore==0.17.3
httpx==0.24.1
huggingface-hub==0.16.4
idna==3.4
importlib-metadata==6.8.0
Jinja2==3.1.2
joblib==1.3.2
langchain==0.0.252
langsmith==0.0.18
llama-index==0.7.22
loguru==0.7.0
MarkupSafe==2.1.3
marshmallow==3.20.1
mpmath==1.3.0
multidict==6.0.4
multiprocess==0.70.15
mypy-extensions==1.0.0
nest-asyncio==1.5.7
networkx==3.1
nltk==3.8.1
nodeenv==1.8.0
numexpr==2.8.4
numpy==1.25.2
openai==0.27.8
openapi-schema-pydantic==1.2.4
opentelemetry-api==1.19.0
opentelemetry-exporter-otlp==1.19.0
opentelemetry-exporter-otlp-proto-grpc==1.19.0
opentelemetry-exporter-otlp-proto-http==1.19.0
opentelemetry-instrumentation==0.40b0
opentelemetry-proto==1.19.0
opentelemetry-sdk==1.19.0
opentelemetry-semantic-conventions==0.40b0
packaging==23.1
pandas==2.0.3
pdfminer.six==20221105
pdfplumber==0.10.2
Pillow==10.0.0
pinecone-client==2.2.2
pip==23.2.1
prisma==0.9.1
protobuf==4.23.4
psutil==5.9.5
pyarrow==12.0.1
pycparser==2.21
pydantic==1.10.12
PyJWT==2.8.0
pypdf==3.14.0
PyPDF2==3.0.1
pypdfium2==4.18.0
python-dateutil==2.8.2
python-dotenv==1.0.0
python-engineio==4.5.1
python-graphql-client==0.4.3
python-socketio==5.8.0
pytz==2023.3
PyYAML==6.0.1
regex==2023.6.3
requests==2.31.0
responses==0.18.0
safetensors==0.3.1
scikit-learn==1.3.0
scipy==1.11.1
seqeval==1.2.2
setuptools==56.0.0
six==1.16.0
sniffio==1.3.0
soupsieve==2.4.1
span-marker==1.2.4
SQLAlchemy==2.0.19
starlette==0.27.0
sympy==1.12
syncer==2.0.3
tenacity==8.2.2
threadpoolctl==3.2.0
tiktoken==0.4.0
tokenizers==0.13.3
tomli==2.0.1
tomlkit==0.12.1
torch==2.0.1
tqdm==4.65.0
transformers==4.31.0
typing_extensions==4.7.1
typing-inspect==0.9.0
tzdata==2023.3
uptrace==1.19.0
urllib3==1.26.16
uvicorn==0.22.0
watchfiles==0.19.0
websockets==11.0.3
win32-setctime==1.1.0
wrapt==1.15.0
xxhash==3.3.0
yarl==1.9.2
zipp==3.16.2
